+++ Navigation/Profile/TextFieldWithPadding.swift
@@ -26,4 +26,3 @@ class TextFieldWithPadding: UITextField {
     }
 }
 
-
